requirements.txt
# =============================================================================
# CV Extractor - Python Dependencies
# =============================================================================

# Web Framework
flask>=2.3.0
werkzeug>=2.3.0

# Environment & Configuration
python-dotenv>=1.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdf2image>=1.16.0
pillow>=10.0.0

# AI & API Integration
google-generativeai>=0.4.0
requests>=2.31.0

# Data Processing & Analysis
numpy>=1.24.0
pandas>=2.0.0

# Testing & Development
pytest>=7.0.0
coverage>=7.0.0

# System Utilities
psutil>=5.9.0
tqdm>=4.66.0